move Router to IbcRouter
