+++ Cargo.toml
@@ -26,7 +26,7 @@ cosmrs = "0.14.0"
 displaydoc = { version = "0.2", default-features = false }
 derive_more = { version = "0.99.17", default-features = false, features = ["from", "into", "display"] }
 ed25519 = { version = "2.1.0", default-features = false }
-ibc = { git = "https://github.com/cosmos/ibc-rs.git", rev= "c20e491" }
+ibc = { git = "https://github.com/cosmos/ibc-rs.git", rev= "11fa19e" }
 ibc-proto = { version = "0.32", default-features = false, features = ["server"] }
 ics23 = { version = "0.10.1", default-features = false, features = ["host-functions"] }
 primitive-types = { version = "0.12.0", default-features = false, features = ["serde_no_std"] }
